requirements.txt
Django==2.1.7
django-classy-tags==0.8.0
django-crispy-forms==1.7.2
django-taggit==0.23.0
django-taggit-templatetags2==1.6.1
pep8==1.7.1
pytz==2018.9
